3459번 승자예측하기 이진 트리를 그려보면 홀수 depth에는 Alice가 짝수 depth에는 Bob이 고를 수 있는 숫자이다. 승자가 누군지 가려보면, 2-5(Alice) 6-9(Bob) 10-25(Alice) 26-41(Bob) 42-105(Alice) 즉, N=1일때를 제외하고 Alice4번-Bob4번-Alice16번-Bob16번-Alice64번-Bob64번.... 이런 패턴. 패턴을 찾긴 했지만.. 더 똑똑한 방법은 못찾겠음...

+++ SWExpertAcademy/[3459]승자예측하기.cpp
@@ -11,10 +11,10 @@ int main() {
     cin >> Testcase;
 
     for (int t = 1; t <= Testcase; t++) {
-        int inputN;
+        long long inputN;
         cin >> inputN;
 
-        int sub = 4;
+        long long sub = 4;
         inputN -= 2;
         string answer = "";
